openstack-setup/requirements.txt
appdirs==1.4.0
Babel==2.3.4
cliff==2.2.0
cmd2==0.6.9
debtcollector==1.9.0
funcsigs==1.0.2
functools32==3.2.3.post2
iso8601==0.1.11
jsonpatch==1.14
jsonpointer==1.10
jsonschema==2.5.1
keystoneauth1==2.15.0
monotonic==1.2
msgpack-python==0.4.8
netaddr==0.7.18
netifaces==0.10.5
openstacksdk==0.9.9
os-client-config==1.22.0
osc-lib==1.2.0
oslo.config==3.19.0
oslo.i18n==3.10.0
oslo.serialization==2.14.0
oslo.utils==3.18.0
pbr==1.10.0
positional==1.1.1
prettytable==0.7.2
pyparsing==2.1.10
python-cinderclient==1.9.0
python-glanceclient==2.5.0
python-keystoneclient==3.6.0
python-neutronclient==6.0.0
python-novaclient==6.0.0
python-openstackclient==3.3.0
pytz==2016.7
PyYAML==3.12
requests==2.11.1
requestsexceptions==1.1.3
rfc3986==0.4.1
simplejson==3.10.0
six==1.10.0
stevedore==1.18.0
unicodecsv==0.14.1
warlock==1.2.0
wrapt==1.10.8
